refactor shared vector-shaped combinator code
