+++ src/combinators/combinators.jl
@@ -1,7 +1,9 @@
+# code shared by vector-shaped combinators
+include("vector.jl")
+
 # built-in generative function combinators
-include("vector_trace.jl")
-#include("recurse/recurse.jl")
-include("map/map.jl")
 include("choice_at/choice_at.jl")
 include("call_at/call_at.jl")
-#include("unfold/unfold.jl")
+include("map/map.jl")
+include("unfold/unfold.jl")
+#include("recurse/recurse.jl")
