+++ src/combinators/map/force_update.jl
@@ -78,9 +78,9 @@ function force_update(args::Tuple, argdiff, trace::VectorTrace{MapType,T,U},
     retained_and_constrained = get_retained_and_constrained(assmt, prev_length, new_length)
 
     # handle removed applications
-    (discard, num_nonempty, score_decrement, noise_decrement) = map_force_update_delete(
+    (discard, num_nonempty, score_decrement, noise_decrement) = vector_force_update_delete(
         new_length, prev_length, trace)
-    (subtraces, retval) = map_remove_deleted_applications(
+    (subtraces, retval) = vector_remove_deleted_applications(
         trace.subtraces, trace.retval, prev_length, new_length)
     score = trace.score - score_decrement
     noise = trace.noise - noise_decrement
@@ -94,7 +94,7 @@ function force_update(args::Tuple, argdiff, trace::VectorTrace{MapType,T,U},
     process_all_new!(gen_fn, args, assmt, prev_length, new_length, state)
 
     # retdiff
-    retdiff = compute_retdiff(state.isdiff_retdiffs, new_length, prev_length)
+    retdiff = vector_compute_retdiff(state.isdiff_retdiffs, new_length, prev_length)
 
     # new trace
     new_trace = VectorTrace{MapType,T,U}(gen_fn, state.subtraces, state.retval, args,
